Refactor derive
