+++ burn-core/src/module/param/mod.rs
@@ -1,7 +1,6 @@
 mod base;
 mod constant;
 mod id;
-mod module;
 mod primitive;
 mod running;
 mod tensor;
@@ -9,7 +8,6 @@ mod visitor;
 
 pub use base::*;
 pub use id::*;
-pub use module::*;
 pub use running::*;
 pub use tensor::*;
 pub use visitor::*;
