+++ burn-derive/src/module/base.rs
@@ -1,4 +1,4 @@
-use super::param::Param;
+use super::fn_generator::FnGenerator;
 use crate::module::display;
 use proc_macro::TokenStream;
 use quote::quote;
@@ -9,18 +9,18 @@ pub(crate) fn module_derive_impl(ast: &syn::DeriveInput) -> TokenStream {
 
     let display_fn = display::display_fn(name);
 
-    let param = Param::from_ast(ast);
-    let num_params_fn = param.gen_num_params_fn();
-    let visit = param.gen_visit_fn();
-    let map_mut = param.gen_map_fn();
-    let devices_fn = param.gen_devices_fn();
-    let to_device_fn = param.gen_to_device_fn();
-    let state_fn = param.gen_state_fn();
-    let load_fn = param.gen_load_fn();
-    let inner_fn = param.gen_inner_fn();
-    let from_inner_fn = param.gen_from_inner_fn();
-    let detach_fn = param.gen_detach_fn();
-    let clone_fn = param.gen_clone_fn();
+    let generator = FnGenerator::from_ast(ast);
+    let num_params_fn = generator.gen_num_params_fn();
+    let visit = generator.gen_visit_fn();
+    let map_mut = generator.gen_map_fn();
+    let devices_fn = generator.gen_devices_fn();
+    let to_device_fn = generator.gen_to_device_fn();
+    let state_fn = generator.gen_state_fn();
+    let load_fn = generator.gen_load_fn();
+    let inner_fn = generator.gen_inner_fn();
+    let from_inner_fn = generator.gen_from_inner_fn();
+    let detach_fn = generator.gen_detach_fn();
+    let clone_fn = generator.gen_clone_fn();
     let generics_names_except_backend = generics_names_except_backend(&ast.generics);
 
     let gen = quote! {
