+++ burn-derive/src/module/mod.rs
@@ -1,5 +1,5 @@
 pub(crate) mod display;
-pub(crate) mod param;
+pub(crate) mod fn_generator;
 
 mod base;
